Decompile aiShowText

+++ src/include/game/game_0601b0.h
@@ -558,7 +558,7 @@ u32 menuhandler001a6d4c(u32, u32, u32);
 
 struct defaultobj *objFindByTagId(s32 tag_id);
 
-char *textGet(u16 textid);
+char *textGet(u32 textid);
 
 u32 getDifficulty(void);
